Read from job streams and write to Output.log
