+++ source/Octopus.Tentacle/Scripts/Kubernetes/RunningKubernetesJobScript.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using k8s;
 using k8s.Models;
 using Nito.AsyncEx;
 using Octopus.Diagnostics;
@@ -30,6 +31,7 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
         readonly CancellationToken scriptCancellationToken;
         readonly string? instanceName;
         readonly AsyncLazy<string> bootstrapRunnerScript;
+        readonly KubernetesJobOutputStreamWriter outputStreamWriter;
 
         public int ExitCode { get; private set; }
         public ProcessState State { get; private set; }
@@ -66,6 +68,7 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
                 using var reader = new StreamReader(stream!, Encoding.UTF8);
                 return await reader.ReadToEndAsync();
             });
+            outputStreamWriter = new KubernetesJobOutputStreamWriter(workspace);
         }
 
         public async Task Execute(CancellationToken taskCancellationToken)
@@ -132,13 +135,13 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
 
             var checkJobTask = CheckIfJobHasCompleted(cancellationToken, jobCompletionCancellationTokenSource);
 
-            //we pass the job completion CTS here because
-            var monitorJobOutputTask = ReadJobOutputStreams(writer, jobCompletionCancellationTokenSource.Token);
+            //we pass the job completion CTS here because its used to cancel the writing of the job stream
+            var monitorJobOutputTask = outputStreamWriter.StreamJobLogsToScriptLog(writer, jobCompletionCancellationTokenSource.Token);
 
             await Task.WhenAll(checkJobTask, monitorJobOutputTask);
 
-            //once they have both finished, perform one last log read
-            await ReadJobOutputStreams(writer, cancellationToken);
+            //once they have both finished, perform one last log read (and don't cancel on it)
+            await outputStreamWriter.StreamJobLogsToScriptLog(writer, CancellationToken.None, true);
 
             //return the exit code of the jobs
             return checkJobTask.Result;
@@ -173,86 +176,7 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
             return resultStatusCode;
         }
 
-        async Task ReadJobOutputStreams(IScriptLogWriter writer, CancellationToken cancellationToken)
-        {
-            //open the file streams for reading
-            using var stdOutStream = new StreamReader(workspace.OpenFileStreamForReading("stdout.log"), Encoding.UTF8);
-            using var stdErrStream = new StreamReader(workspace.OpenFileStreamForReading("stderr.log"), Encoding.UTF8);
-
-            long lastStdOutOffset = 0;
-            long lastStdErrOffset = 0;
-
-            while (true)
-            {
-                if (cancellationToken.IsCancellationRequested)
-                    break;
-
-                //Read both the stdout and stderr log files
-                var stdOutReadTask = ReadLogFileTail(writer, stdOutStream, ProcessOutputSource.StdOut, lastStdOutOffset);
-                var stdErrReadTask = ReadLogFileTail(writer, stdErrStream, ProcessOutputSource.StdErr, lastStdErrOffset);
-
-                //wait for them to both complete
-                await Task.WhenAll(stdOutReadTask, stdErrReadTask);
-
-                //store the final offsets
-                lastStdOutOffset = stdOutReadTask.Result.FinalOffset;
-                lastStdErrOffset = stdErrReadTask.Result.FinalOffset;
-
-                //stitch the log lines together and order by occurred, then write to actual log
-                var orderedLogLines = stdOutReadTask.Result.Logs
-                    .Concat(stdErrReadTask.Result.Logs)
-                    .OrderBy(ll => ll.Occurred);
-
-                //write all the read log lines to the output script log
-                foreach (var logLine in orderedLogLines)
-                {
-                    writer.WriteOutput(logLine.Source, logLine.Message, logLine.Occurred);
-                }
-
-                //wait for 250ms before reading the logs again
-                await Task.Delay(TimeSpan.FromMilliseconds(250), cancellationToken);
-            }
-        }
-
-        record LogLine(ProcessOutputSource Source, string Message, DateTimeOffset Occurred);
-
-        static async Task<(IEnumerable<LogLine> Logs, long FinalOffset)> ReadLogFileTail(IScriptLogWriter writer, StreamReader reader, ProcessOutputSource source, long lastOffset)
-        {
-            if (reader.BaseStream.Length == lastOffset)
-                return (Enumerable.Empty<LogLine>(), lastOffset);
-
-            reader.BaseStream.Seek(lastOffset, SeekOrigin.Begin);
-
-            var newLines = new List<LogLine>();
-            string? line;
-            do
-            {
-                line = await reader.ReadLineAsync();
-                if (line.IsNullOrEmpty())
-                    break;
-
-                var logParts = line!.Split(new[] { '|' }, 2);
-
-                if (logParts.Length != 2)
-                {
-                    writer.WriteOutput(ProcessOutputSource.StdErr, $"Invalid log line detected. '{line}' is not correctly pipe-delimited.");
-                    continue;
-                }
-
-                //part 1 is the datetimeoffset
-                if (!DateTimeOffset.TryParse(logParts[0], out var occurred))
-                {
-                    writer.WriteOutput(ProcessOutputSource.StdErr, $"Failed to parse '{logParts[0]}' as a DateTimeOffset. Using DateTimeOffset.UtcNow.");
-                    occurred = DateTimeOffset.UtcNow;
-                    ;
-                }
-
-                //add the new line
-                newLines.Add(new LogLine(source, logParts[1], occurred));
-            } while (!line.IsNullOrEmpty());
-
-            return (newLines, reader.BaseStream.Position);
-        }
+        
 
         async Task CreateJob(IScriptLogWriter writer, CancellationToken cancellationToken)
         {
@@ -264,8 +188,8 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
             var jobName = jobService.BuildJobName(scriptTicket);
 
             //Deserialize the volume configuration
-            var volumes = k8s.KubernetesYaml.Deserialize<List<V1Volume>>(KubernetesConfig.JobVolumeYaml);
-            
+            var volumes = KubernetesYaml.Deserialize<List<V1Volume>>(KubernetesConfig.JobVolumeYaml);
+
             var job = new V1Job
             {
                 ApiVersion = "batch/v1",
@@ -296,9 +220,9 @@ namespace Octopus.Tentacle.Scripts.Kubernetes
                                     Command = new List<string> { "bash" },
                                     Args = new List<string>
                                     {
-                                        $"'/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}/bootstrapRunner.sh'",
-                                        $"'/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}'",
-                                        $"'/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}/{scriptName}'"
+                                        $"/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}/bootstrapRunner.sh",
+                                        $"/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}",
+                                        $"/data/tentacle-home/{instanceName}/Work/{scriptTicket.TaskId}/{scriptName}"
                                     }.Concat((workspace.ScriptArguments ?? Array.Empty<string>())
                                         .SelectMany(arg => new[]
                                         {
